* commit 'e52a5bca29f5c9563ac9dab2dbeb4220fa37d970':
  Fix website due to arithmetic overflow

+++ Gateway.Web/Database/ResponseStats.cs
@@ -22,7 +22,7 @@ namespace Gateway.Web.Database
             }
         }
 
-        private void Add(string controller, int? resultCode, int? count, int? avgResponseMs)
+        private void Add(string controller, int? resultCode, int? count, long? avgResponseMs)
         {
             ResponseStat stat;
             if (!_stats.TryGetValue(controller, out stat))
@@ -84,6 +84,6 @@ namespace Gateway.Web.Database
     {
         public int SuccessCount { get; set; }
         public int FailureCount { get; set; }
-        public int AvgResponseMs { get; set; }
+        public long AvgResponseMs { get; set; }
     }
 }
